refactor: 아티스트 페이지 구조 정리
