+++ apps/client/app/[exhibitionId]/artist/_components/ArtistPageClient.tsx
@@ -5,13 +5,12 @@ import { Divider } from "@/components/common/Divider/Divider";
 import { ScrollTabBar } from "@/components/common/ScrollTabBar/ScrollTabBar";
 import { useScrollSpy } from "@/components/common/ScrollTabBar/useScrollSpy";
 import { Header } from "../../_components/Header";
-import type { ArtistProfile } from "../api/artist";
-import type { PartnerPart } from "../api/partner";
+import type { ArtistProfile } from "@/lib/api/artists/artist";
+import type { PartnerPart } from "@/lib/api/partner";
 import { ArtistSection } from "./sections/ArtistSection";
 import { PartnerSection } from "./sections/PartnerSection";
 
 interface ArtistPageClientProps {
-	schoolId: string;
 	exhibitionId: string;
 	partners: PartnerPart[];
 	artists: ArtistProfile[];
