+++ apps/client/app/[exhibitionId]/artist/_components/sections/ArtistSection.tsx
@@ -1,7 +1,6 @@
-import { useMemo } from "react";
 import { CardGrid } from "@/components/common/Card/CardGrid";
 import { Title } from "@/components/common/Title/Title";
-import type { ArtistProfile } from "../../api/artist";
+import type { ArtistProfile } from "@/lib/api/artists/artist";
 
 interface ArtistSectionProps {
 	exhibitionId: string;
@@ -9,21 +8,15 @@ interface ArtistSectionProps {
 }
 
 export function ArtistSection({ exhibitionId, artists = [] }: ArtistSectionProps) {
-	const cardItems = useMemo(() => {
-		const mapped = artists.map((artist) => ({
-			id: artist.profileId,
-			title: artist.nameKo || "이름",
-			author: artist.nameEn || "Name",
-			imageUrl: artist.profileImg,
-			category: "",
-		}));
+	if (!artists.length) return null;
 
-		return mapped;
-	}, [artists]);
-
-	if (!artists.length) {
-		return null;
-	}
+	const cardItems = artists.map((artist) => ({
+		id: artist.profileId,
+		title: artist.nameKo || "이름",
+		author: artist.nameEn || "Name",
+		imageUrl: artist.profileImg,
+		category: "",
+	}));
 
 	return (
 		<section className="flex flex-col pt-4 pb-6">
